Moved the JPro-App to an own subproject
