+++ subprojects/sampler-jpro/sampler-jpro.gradle
@@ -21,11 +21,10 @@
  */
 
 apply plugin: 'application'
+apply plugin: 'com.sandec.jpro'
 
-mainClassName = 'org.kordamp.bootstrapfx.Sampler'
+mainClassName = 'org.kordamp.bootstrapfx.SamplerJPro'
 
 dependencies {
-    compile project(':bootstrapfx-core')
-    compile 'org.fxmisc.richtext:richtextfx:0.9.1'
-    compile group: 'org.apache.commons', name: 'commons-io', version: '1.3.2'
+    compile project(':sampler')
 }
