+++ subprojects/sampler/src/main/java/org/kordamp/bootstrapfx/Sampler.java
@@ -22,89 +22,25 @@
 package org.kordamp.bootstrapfx;
 
 import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.geometry.Side;
-import javafx.scene.Node;
 import javafx.scene.Scene;
-import javafx.scene.control.Tab;
-import javafx.scene.control.TabPane;
-import javafx.scene.layout.StackPane;
-import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
-
-import java.net.URL;
-import java.nio.file.Paths;
-import org.apache.commons.io.IOUtils;
-
-import static java.nio.file.Files.lines;
-import static java.util.stream.Collectors.joining;
 
 public class Sampler extends Application {
     @Override
     public void start(Stage primaryStage) throws Exception {
-        URL location = getClass().getResource("sampler.fxml");
-        FXMLLoader fxmlLoader = new FXMLLoader(location);
-        TabPane tabPane = fxmlLoader.load();
 
-        tabPane.getTabs().add(new DemoTab("Buttons", "buttons.fxml"));
-        tabPane.getTabs().add(new DemoTab("Labels", "labels.fxml"));
-        tabPane.getTabs().add(new DemoTab("Alerts", "alerts.fxml"));
-        tabPane.getTabs().add(new DemoTab("Panels", "panels.fxml"));
-        tabPane.getTabs().add(new DemoTab("Headings", "text.fxml"));
-        tabPane.getTabs().add(new DemoTab("Progress Bars", "progressbars.fxml"));
-        tabPane.getTabs().add(new DemoTab("Tooltips ", "tooltips.fxml"));
-        tabPane.getTabs().add(new DemoTab("Text ", "text2.fxml"));
-        tabPane.getTabs().add(new DemoTab("Paragraph ", "paragraph.fxml"));
-        tabPane.getTabs().add(new DemoTab("Button Groups ", "button_groups.fxml"));
-        tabPane.getTabs().add(new DemoTab("SplitMenuButtons", "split_menu_buttons.fxml"));
-
-        StackPane innerPane = new StackPane(tabPane);
-        innerPane.setMaxWidth(1000);
-        innerPane.setMaxHeight(600);
-        innerPane.setStyle("-fx-background-color: #ddd; -fx-background-radius: 10;");
-        StackPane outerPane = new StackPane(innerPane);
-        outerPane.setStyle("-fx-background-image: url('/org/kordamp/bootstrapfx/ambient-background.jpg');" +
-                "-fx-background-size: cover;");
-
-        Scene scene = new Scene(outerPane);
+        Scene scene = new Scene(new DemoTabPane());
         scene.getStylesheets().addAll(
-            "org/kordamp/bootstrapfx/bootstrapfx.css",
-            "org/kordamp/bootstrapfx/sampler.css",
-            "org/kordamp/bootstrapfx/xml-highlighting.css");
+                "bootstrapfx.css",
+                "org/kordamp/bootstrapfx/sampler.css",
+                "org/kordamp/bootstrapfx/xml-highlighting.css");
 
         primaryStage.setTitle("BootstrapFX Sampler");
         primaryStage.setScene(scene);
         //primaryStage.sizeToScene();
         primaryStage.setWidth(1024);
         primaryStage.show();
+
     }
 
-    private static class DemoTab extends Tab {
-        private DemoTab(String title, String sourceFile) throws Exception {
-            super(title);
-            setClosable(false);
-
-            TabPane content = new TabPane();
-            setContent(content);
-            content.setSide(Side.BOTTOM);
-
-            Tab widgets = new Tab("Widgets");
-            widgets.setClosable(false);
-            URL location = getClass().getResource(sourceFile);
-            FXMLLoader fxmlLoader = new FXMLLoader(location);
-            Node node = fxmlLoader.load();
-            widgets.setContent(node);
-
-            Tab source = new Tab("Source");
-            source.setClosable(false);
-            XMLEditor editor = new XMLEditor();
-            editor.setEditable(false);
-
-            String text = IOUtils.toString(getClass().getResourceAsStream(sourceFile));
-            editor.setText(text);
-            source.setContent(editor);
-
-            content.getTabs().addAll(widgets, source);
-        }
-    }
 }
